+++ integration/type-registry/foo.ts
@@ -3,6 +3,7 @@ import { messageTypeRegistry } from './typeRegistry';
 import { util, configure, Writer, Reader } from 'protobufjs/minimal';
 import * as Long from 'long';
 import { Timestamp } from './google/protobuf/timestamp';
+import { Struct } from './google/protobuf/struct';
 
 export const protobufPackage = 'foo';
 
@@ -14,6 +15,11 @@ export interface Foo {
 export interface Foo2 {
   $type: 'foo.Foo2';
   timestamp: Date | undefined;
+}
+
+export interface WithStruct {
+  $type: 'foo.WithStruct';
+  struct: { [key: string]: any } | undefined;
 }
 
 function createBaseFoo(): Foo {
@@ -124,6 +130,60 @@ export const Foo2 = {
 
 messageTypeRegistry.set(Foo2.$type, Foo2);
 
+function createBaseWithStruct(): WithStruct {
+  return { $type: 'foo.WithStruct', struct: undefined };
+}
+
+export const WithStruct = {
+  $type: 'foo.WithStruct' as const,
+
+  encode(message: WithStruct, writer: Writer = Writer.create()): Writer {
+    if (message.struct !== undefined) {
+      Struct.encode(Struct.wrap(message.struct), writer.uint32(10).fork()).ldelim();
+    }
+    return writer;
+  },
+
+  decode(input: Reader | Uint8Array, length?: number): WithStruct {
+    const reader = input instanceof Reader ? input : new Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseWithStruct();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.struct = Struct.unwrap(Struct.decode(reader, reader.uint32()));
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): WithStruct {
+    return {
+      $type: WithStruct.$type,
+      struct: isObject(object.struct) ? object.struct : undefined,
+    };
+  },
+
+  toJSON(message: WithStruct): unknown {
+    const obj: any = {};
+    message.struct !== undefined && (obj.struct = message.struct);
+    return obj;
+  },
+
+  fromPartial<I extends Exact<DeepPartial<WithStruct>, I>>(object: I): WithStruct {
+    const message = createBaseWithStruct();
+    message.struct = object.struct ?? undefined;
+    return message;
+  },
+};
+
+messageTypeRegistry.set(WithStruct.$type, WithStruct);
+
 type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;
 
 export type DeepPartial<T> = T extends Builtin
@@ -168,6 +228,10 @@ function fromJsonTimestamp(o: any): Date {
 if (util.Long !== Long) {
   util.Long = Long as any;
   configure();
+}
+
+function isObject(value: any): boolean {
+  return typeof value === 'object' && value !== null;
 }
 
 function isSet(value: any): boolean {