+++ integration/type-registry/google/protobuf/struct.ts
@@ -1,4 +1,5 @@
 /* eslint-disable */
+import { messageTypeRegistry } from '../../typeRegistry';
 import { util, configure, Writer, Reader } from 'protobufjs/minimal';
 import * as Long from 'long';
 
@@ -12,8 +13,8 @@ export const protobufPackage = 'google.protobuf';
  */
 export enum NullValue {
   /** NULL_VALUE - Null value. */
-  NULL_VALUE = 'NULL_VALUE',
-  UNRECOGNIZED = 'UNRECOGNIZED',
+  NULL_VALUE = 0,
+  UNRECOGNIZED = -1,
 }
 
 export function nullValueFromJSON(object: any): NullValue {
@@ -37,15 +38,6 @@ export function nullValueToJSON(object: NullValue): string {
   }
 }
 
-export function nullValueToNumber(object: NullValue): number {
-  switch (object) {
-    case NullValue.NULL_VALUE:
-      return 0;
-    default:
-      return 0;
-  }
-}
-
 /**
  * `Struct` represents a structured data value, consisting of fields
  * which map to dynamically typed values. In some languages, `Struct`
@@ -57,11 +49,13 @@ export function nullValueToNumber(object: NullValue): number {
  * The JSON representation for `Struct` is JSON object.
  */
 export interface Struct {
+  $type: 'google.protobuf.Struct';
   /** Unordered map of dynamically typed values. */
   fields: { [key: string]: any | undefined };
 }
 
 export interface Struct_FieldsEntry {
+  $type: 'google.protobuf.Struct.FieldsEntry';
   key: string;
   value: any | undefined;
 }
@@ -75,6 +69,7 @@ export interface Struct_FieldsEntry {
  * The JSON representation for `Value` is JSON value.
  */
 export interface Value {
+  $type: 'google.protobuf.Value';
   /** Represents a null value. */
   nullValue: NullValue | undefined;
   /** Represents a double value. */
@@ -95,19 +90,25 @@ export interface Value {
  * The JSON representation for `ListValue` is JSON array.
  */
 export interface ListValue {
+  $type: 'google.protobuf.ListValue';
   /** Repeated field of dynamically typed values. */
   values: any[];
 }
 
 function createBaseStruct(): Struct {
-  return { fields: {} };
+  return { $type: 'google.protobuf.Struct', fields: {} };
 }
 
 export const Struct = {
+  $type: 'google.protobuf.Struct' as const,
+
   encode(message: Struct, writer: Writer = Writer.create()): Writer {
     Object.entries(message.fields).forEach(([key, value]) => {
       if (value !== undefined) {
-        Struct_FieldsEntry.encode({ key: key as any, value }, writer.uint32(10).fork()).ldelim();
+        Struct_FieldsEntry.encode(
+          { $type: 'google.protobuf.Struct.FieldsEntry', key: key as any, value },
+          writer.uint32(10).fork()
+        ).ldelim();
       }
     });
     return writer;
@@ -136,6 +137,7 @@ export const Struct = {
 
   fromJSON(object: any): Struct {
     return {
+      $type: Struct.$type,
       fields: isObject(object.fields)
         ? Object.entries(object.fields).reduce<{ [key: string]: any | undefined }>((acc, [key, value]) => {
             acc[key] = value as any | undefined;
@@ -189,11 +191,15 @@ export const Struct = {
   },
 };
 
+messageTypeRegistry.set(Struct.$type, Struct);
+
 function createBaseStruct_FieldsEntry(): Struct_FieldsEntry {
-  return { key: '', value: undefined };
+  return { $type: 'google.protobuf.Struct.FieldsEntry', key: '', value: undefined };
 }
 
 export const Struct_FieldsEntry = {
+  $type: 'google.protobuf.Struct.FieldsEntry' as const,
+
   encode(message: Struct_FieldsEntry, writer: Writer = Writer.create()): Writer {
     if (message.key !== '') {
       writer.uint32(10).string(message.key);
@@ -227,6 +233,7 @@ export const Struct_FieldsEntry = {
 
   fromJSON(object: any): Struct_FieldsEntry {
     return {
+      $type: Struct_FieldsEntry.$type,
       key: isSet(object.key) ? String(object.key) : '',
       value: isSet(object?.value) ? object.value : undefined,
     };
@@ -247,8 +254,11 @@ export const Struct_FieldsEntry = {
   },
 };
 
+messageTypeRegistry.set(Struct_FieldsEntry.$type, Struct_FieldsEntry);
+
 function createBaseValue(): Value {
   return {
+    $type: 'google.protobuf.Value',
     nullValue: undefined,
     numberValue: undefined,
     stringValue: undefined,
@@ -259,9 +269,11 @@ function createBaseValue(): Value {
 }
 
 export const Value = {
+  $type: 'google.protobuf.Value' as const,
+
   encode(message: Value, writer: Writer = Writer.create()): Writer {
     if (message.nullValue !== undefined) {
-      writer.uint32(8).int32(nullValueToNumber(message.nullValue));
+      writer.uint32(8).int32(message.nullValue);
     }
     if (message.numberValue !== undefined) {
       writer.uint32(17).double(message.numberValue);
@@ -289,7 +301,7 @@ export const Value = {
       const tag = reader.uint32();
       switch (tag >>> 3) {
         case 1:
-          message.nullValue = nullValueFromJSON(reader.int32());
+          message.nullValue = reader.int32() as any;
           break;
         case 2:
           message.numberValue = reader.double();
@@ -316,6 +328,7 @@ export const Value = {
 
   fromJSON(object: any): Value {
     return {
+      $type: Value.$type,
       nullValue: isSet(object.nullValue) ? nullValueFromJSON(object.nullValue) : undefined,
       numberValue: isSet(object.numberValue) ? Number(object.numberValue) : undefined,
       stringValue: isSet(object.stringValue) ? String(object.stringValue) : undefined,
@@ -388,11 +401,15 @@ export const Value = {
   },
 };
 
+messageTypeRegistry.set(Value.$type, Value);
+
 function createBaseListValue(): ListValue {
-  return { values: [] };
+  return { $type: 'google.protobuf.ListValue', values: [] };
 }
 
 export const ListValue = {
+  $type: 'google.protobuf.ListValue' as const,
+
   encode(message: ListValue, writer: Writer = Writer.create()): Writer {
     for (const v of message.values) {
       Value.encode(Value.wrap(v!), writer.uint32(10).fork()).ldelim();
@@ -420,6 +437,7 @@ export const ListValue = {
 
   fromJSON(object: any): ListValue {
     return {
+      $type: ListValue.$type,
       values: Array.isArray(object?.values) ? [...object.values] : [],
     };
   },
@@ -453,6 +471,8 @@ export const ListValue = {
   },
 };
 
+messageTypeRegistry.set(ListValue.$type, ListValue);
+
 type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;
 
 export type DeepPartial<T> = T extends Builtin
@@ -462,13 +482,13 @@ export type DeepPartial<T> = T extends Builtin
   : T extends ReadonlyArray<infer U>
   ? ReadonlyArray<DeepPartial<U>>
   : T extends {}
-  ? { [K in keyof T]?: DeepPartial<T[K]> }
+  ? { [K in Exclude<keyof T, '$type'>]?: DeepPartial<T[K]> }
   : Partial<T>;
 
 type KeysOfUnion<T> = T extends T ? keyof T : never;
 export type Exact<P, I extends P> = P extends Builtin
   ? P
-  : P & { [K in keyof P]: Exact<P[K], I[K]> } & Record<Exclude<keyof I, KeysOfUnion<P>>, never>;
+  : P & { [K in keyof P]: Exact<P[K], I[K]> } & Record<Exclude<keyof I, KeysOfUnion<P> | '$type'>, never>;
 
 // If you get a compile-error about 'Constructor<Long> and ... have no overlap',
 // add '--ts_proto_opt=esModuleInterop=true' as a flag when calling 'protoc'.