+++ protos/build.sh
@@ -6,7 +6,7 @@
 protoc \
   --plugin=./node_modules/ts-proto/protoc-gen-ts_proto \
   --ts_proto_out=. \
-  --ts_proto_opt=exportCommonSymbols=false \
+  --ts_proto_opt=exportCommonSymbols=false,unknownFields=true,usePrototypeForDefaults=true \
   ./google/protobuf/descriptor.proto \
   ./google/protobuf/compiler/plugin.proto
 
@@ -14,7 +14,7 @@ protoc \
   ./index.ts \
   ./google/protobuf/descriptor.ts \
   ./google/protobuf/compiler/plugin.ts \
-  --outDir dist --declaration
+  --outDir dist --declaration --downlevelIteration